+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - XML documentation on interfaces
 - Test project with unit tests and end-to-end service container registration tests
+- Dependency Injection component for service registration in the service container
+- `IDependencyInjectionRootComponent` extension `AddMapping` to attach component to root component
 
 ## [0.1.0] - 2020-06-20